requirements.txt
# Core RAG Framework - Strands (Day 1)
strands-agents  # Modern agent framework (replaces LangChain)
strands-agents-tools  # Optional: Pre-built tools for experimentation

# Vector Database
chromadb>=1.4.0

# LLM Provider - OpenRouter
openai>=1.58.1  # OpenRouter uses OpenAI-compatible API

# Progress Bars & Utilities
tqdm>=4.67.1
python-dotenv>=1.0.1

# Optional: Evaluation (for Day 2+)
# ragas>=0.2.6
# datasets>=3.1.0

# Optional: Experiment Tracking (for Day 3)
# wandb>=0.18.7  # Uncomment if using Weights & Biases
# mlflow>=2.19.0  # Uncomment if using MLflow

# Document Loaders
pypdf>=5.1.0
# python-docx>=1.1.2
# openpyxl>=3.1.5

# Chunking Evaluation (for ClusterSemanticChunker)
git+https://github.com/brandonstarxel/chunking_evaluation.git

# Optional: Data Processing (if needed)
# numpy>=1.26.0
# pandas>=2.2.3

# Optional: Visualization (for Day 3)
# matplotlib>=3.9.3
# seaborn>=0.13.2
# plotly>=5.24.1

# API & Dashboard
fastapi>=0.115.6
uvicorn>=0.34.0
streamlit>=1.40.2
requests>=2.31.0

# Evaluation and Observability
ragas>=0.2.6
mlflow>=2.19.0
datasets>=3.1.0

# Guardrails
pyyaml>=6.0.1

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
# jupyter>=1.1.1
# ipykernel>=6.29.5